sonar:cpp:S5276 — make I/O-boundary narrowing conversions explicit
Add static_cast at the socket/byte-parse seams where integer promotion
would otherwise narrow implicitly:

- TCPTransport recv() results (ssize_t->int): safe — each is immediately
  checked 'n <= 0', so a -1 error survives the cast and is handled; a valid
  read of a <=256-byte buffer never overflows int. No negative return masked.
- Big/little-endian CAN-ID and frame-length expressions (int promoted by |
  and <<, narrowed back to uint16_t): SecureTcpTransport, DiscoveryPacket,
  BLESignalSource, VehicleDetector, DBCSignalTranslator.

Behaviour identical (the casts only annotate conversions the compiler
already performed implicitly).

+++ src/discovery/DiscoveryPacket.cpp
@@ -17,7 +17,7 @@ static void writeBigEndian(uint8_t* dst, uint64_t value) {
 }
 
 static uint16_t readBigEndian16(const uint8_t* src) {
-    return (static_cast<uint16_t>(src[0]) << 8) | static_cast<uint16_t>(src[1]);
+    return static_cast<uint16_t>((static_cast<uint16_t>(src[0]) << 8) | static_cast<uint16_t>(src[1]));
 }
 
 static uint64_t readBigEndian64(const uint8_t* src) {
